JWT Authentication Setup

+++ backend/controllers/goalController.js
@@ -47,7 +47,7 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error('User not found');
   }
 
-  // make syre the logged in user matches the goal user
+  // make sure the logged in user matches the goal user
   if (goal.user.toString() !== user.id) {
     res.status(401);
     throw new Error('User not authorized');
@@ -69,6 +69,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   await goal.deleteOne({ id: req.params.id });
